reformat remax's code
