+++ examples/remax_trainer/run.sh
@@ -5,15 +5,17 @@ export TRANSFORMERS_OFFLINE=1
 
 export VLLM_ATTENTION_BACKEND=XFORMERS
 
+DATA_DIR=./data
+
 python3 -m verl.trainer.main_ppo \
     algorithm.adv_estimator=remax \
-    data.train_files=$HOME/data/gsm8k/train.parquet \
-    data.val_files=$HOME/data/gsm8k/train.parquet \
+    data.train_files=$DATA_DIR/gsm8k/train.parquet \
+    data.val_files=$DATA_DIR/gsm8k/test.parquet \
     data.train_batch_size=512 \
     data.val_batch_size=1312 \
     data.max_prompt_length=512 \
     data.max_response_length=1024 \
-    actor_rollout_ref.model.path=Qwen/Qwen2.5-3B-Instruct \
+    actor_rollout_ref.model.path=/220049033/models/Qwen/Qwen2.5-3B-Instruct \
     actor_rollout_ref.actor.optim.lr=1e-6 \
     actor_rollout_ref.model.use_remove_padding=True \
     actor_rollout_ref.actor.ppo_mini_batch_size=128 \
@@ -33,7 +35,7 @@ python3 -m verl.trainer.main_ppo \
     actor_rollout_ref.ref.fsdp_config.param_offload=True \
     algorithm.kl_ctrl.kl_coef=0.001 \
     trainer.critic_warmup=0 \
-    trainer.logger=['console','wandb'] \
+    trainer.logger=['console'] \
     trainer.project_name='verl_remax_example_gsm8k' \
     trainer.experiment_name='qwen2.5_3b_function_rm_kl1e-3' \
     +trainer.val_before_train=False \
